Fix main sect, modify html places, write content
Fix some problems with overflow main image and the flex items of the
main section when wrapping. Wrap the image element in a div and modify
the class selectors of these items for this.

Add item titles in the HTML places section and a new paragraph and
create the class selectors of them in the style.css file.

Write the text content of the main and places sections.

Fix header padding.